Add BoxWriterScript interpreter and JSON
BoxWriterScript.run will run a script against a revision, putting
tokens to a mutable writer (pure writer would also be fine, in theory
stream should be possible?)
Add JSON reader/writer, demo with writer.
Move differentiation of primitive token types - they are now separate
BoxReaderDeltaF cases, to allow them to be dealt with at the
reader/writer level, so that (for example) JSON can perform necessary
safe up/down casts to work with JSON data types, but where a
reader/writer can exactly represent a primitive including type, it may
choose not to perform casts.
Fix New: CacheResult using Int id instead of Long.

+++ src/main/scala/org/rebeam/boxes/persistence/CacheResult.scala
@@ -19,6 +19,6 @@ case class Cached(id: Long) extends CacheResult {
  * Object was not already cached, it is now cached using the provided id
  * @param id The id under which the object has now been cached
  */
-case class New(id:Int) extends CacheResult {
+case class New(id:Long) extends CacheResult {
     def isCached = false
 }
